pyproject.toml
[project]
name = "portfolio-website"
version = "0.1.0"
description = "Static site generator for my portfolio website"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "jinja2>=3.1.6",
    "mistune>=3.1.3",
    "python-frontmatter>=1.1.0",
]
readme = "README.md"

[dependency-groups]
dev = [
    "flask>=3.1.0",
]